ADD : Register and deregister from ride

+++ backend/src/main/java/com/poc/capcovoit/dao/RideRepository.java
@@ -4,6 +4,6 @@ import com.poc.capcovoit.entity.Ride;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RideRepository extends JpaRepository<Ride, Long> {
+public interface RideRepository extends JpaRepository<Ride, Integer> {
     List<Ride> findByDriverEmail(String driverEmail);
 }
